src/monitoring/evidently_service/requirements.txt
dataclasses
Flask
pandas
Werkzeug
requests
prometheus_client
pyyaml
evidently
gunicorn
